Refactor JavaScript files to optimize structure
