+++ src/js/PC/Instructions/ADD.js
@@ -1,7 +1,7 @@
 "use strict";
-import Info from "../../Info";
-import ALU from "../ALU";
-import RAM from "../RAM";
+import Info from "../Parts/Info";
+import ALU from "../Parts/ALU";
+import RAM from "../Parts/RAM";
 
 export default class ADD {
   constructor() {
