+++ src/js/PC/Instructions/JEQ.js
@@ -1,7 +1,6 @@
-"use strict";
-import Info from "../../Info";
-import ALU from "../ALU";
-import RAM from "../RAM";
+import Info from "../Parts/Info";
+import ALU from "../Parts/ALU";
+import RAM from "../Parts/RAM";
 
 export default class JEQ {
   constructor() {
